+++ libs/common/src/lib/interfaces/portfolio-report-rule.interface.ts
@@ -1,5 +1,4 @@
 export interface PortfolioReportRule {
-  categoryName: string;
   configuration?: {
     threshold?: {
       max: number;